refactor: move types to seperate file
